clarifai-backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
feedparser>=6.0.10
requests>=2.31.0
beautifulsoup4>=4.12.0
google-generativeai>=0.4.1
langchain-google-genai>=1.0.0
langchain-neo4j>=0.1.0
langchain-core>=0.1.0
neo4j>=5.15.0
python-dateutil>=2.8.2
aiohttp>=3.9.0
